+++ src/Helper/Product/Stock/Factory.php
@@ -21,9 +21,34 @@
 
 namespace Shopgate\Export\Helper\Product\Stock;
 
-use Shopgate\Export\Model\Shopgate\Product\StockItem;
+use Magento\Framework\Module\Manager;
+use Magento\Framework\ObjectManagerInterface;
 
-interface Utility
+class Factory
 {
-    public function getStockItem($product): StockItem;
+    /** @var Manager */
+    private $moduleManager;
+
+    /** @var ObjectManagerInterface */
+    private $objectManager;
+
+    /**
+     * @param Manager $moduleManager
+     * @param ObjectManagerInterface $objectManager
+     */
+    public function __construct(Manager $moduleManager, ObjectManagerInterface $objectManager)
+    {
+        $this->moduleManager = $moduleManager;
+        $this->objectManager = $objectManager;
+    }
+
+    /**
+     * @return Utility
+     */
+    public function getUtility()
+    {
+        return $this->moduleManager->isEnabled('Magento_InventorySalesApi')
+            ? $this->objectManager->create(UtilityInventorySalesApi::class)
+            : $this->objectManager->create(UtilityCommon::class);
+    }
 }